+++ package.json
@@ -32,7 +32,6 @@
     "dotenv": "^8.2.0",
     "easy-soap-request": "^4.1.3",
     "filesize": "^10.1.4",
-    "formidable": "^3.5.1",
     "http-status-codes": "^2.1.4",
     "jsonwebtoken": "^8.5.1",
     "lbh-frontend": "^3.5.3",
@@ -41,11 +40,9 @@
     "mockdate": "^3.0.2",
     "next": "^12",
     "path": "^0.12.7",
-    "photoswipe": "^5.4.4",
     "qs": "^6.10.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-dropzone": "^14.2.3",
     "react-hook-form": "^6.14.0",
     "react-hotjar": "^5.0.0",
     "react-photo-view": "^1.2.6",
