+++ src/styles/all.scss
@@ -37,3 +37,68 @@ $govuk-font-family: 'GDS Transport', arial, sans-serif !important;
 @import 'components/back-office.scss';
 @import '@reach/dialog/styles.css';
 @import 'components/contact-details-table.scss';
+
+.custom-file-input::file-selector-button {
+    // color: orange;
+
+    // font-weight: bold;
+    // color: dodgerblue;
+    // padding: 0.5em;
+    // border: thin solid grey;
+    // border-radius: 3px;
+
+    width: 0;
+    padding: 0;
+    visibility: hidden;
+}
+
+// .custom-file-input::-webkit-file-upload-button {
+// //   visibility: hidden;
+
+// background: white;
+// }
+.custom-file-input::before {
+  content: 'Choose file(s)';
+
+// content: "";
+  display: inline-block;
+ 
+
+  padding: 13px 22px 11px;
+  // padding: 0.8125rem 1.375rem 0.6875rem;
+  border: 1px solid #00664f;
+  color: #00664f;
+  background: #ffffff;
+  border-radius: 4px;
+  font-size: 19px;
+  font-weight: 400;
+  line-height: 23px;
+
+  cursor: pointer;
+
+
+
+
+//   :focus:not(:active):not(:hover)
+}
+
+.custom-file-input:hover::before {
+    background: #00513f;
+    color: #fff
+  }
+
+
+
+.custom-file-input::focus:not(:active):not(:hover)::before {
+    border-color: #ffc845;
+    color: #0b0c0c;
+    background-color: #ffc845;
+    box-shadow: 0 2px 0 #0b0c0c;
+  }
+
+// .custom-file-input:hover::before {
+//   border-color: black;
+// }
+// .custom-file-input:active::before {
+//   background: -webkit-linear-gradient(top, #e3e3e3, #f9f9f9);
+// }
